Move crash runnable to its own file
